% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_audit.R
\name{plot_fold_balance}
\alias{plot_fold_balance}
\title{Plot fold balance of positives/negatives per fold}
\usage{
plot_fold_balance(fit)
}
\arguments{
\item{fit}{LeakFit}
}
\description{
Plot fold balance of positives/negatives per fold
}
