% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_audit.R
\name{plot_time_acf}
\alias{plot_time_acf}
\title{Plot ACF of test predictions for time-series leakage checks}
\usage{
plot_time_acf(fit, lag.max = 20)
}
\arguments{
\item{fit}{LeakFit}
}
\description{
Plot ACF of test predictions for time-series leakage checks
}
