% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guards.R
\name{predict_guard}
\alias{predict_guard}
\title{Apply a fitted GuardFit transformer to new data}
\usage{
predict_guard(fit, newdata)
}
\arguments{
\item{fit}{GuardFit object returned by .guard_fit()}

\item{newdata}{matrix/data.frame to transform}
}
\description{
Apply a fitted GuardFit transformer to new data
}
