% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_splits.R
\name{show,LeakSplits-method}
\alias{show,LeakSplits-method}
\title{Display summary for LeakSplits objects}
\usage{
\S4method{show}{LeakSplits}(object)
}
\description{
Display summary for LeakSplits objects
}
