% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_audit.R
\name{plot_perm_distribution}
\alias{plot_perm_distribution}
\title{Plot permutation distribution for a LeakAudit object}
\usage{
plot_perm_distribution(audit)
}
\arguments{
\item{audit}{LeakAudit}
}
\description{
Plot permutation distribution for a LeakAudit object
}
